Move version logging function to version utility
